Error Handling

+++ src/Employee.Business/Services/AddressService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using EmployeeApi.Business.Exceptions;
 using EmployeeApi.Business.Validation;
 using EmployeeApi.Common.Dtos.Address;
 using EmployeeApi.Common.Interfaces;
@@ -37,7 +38,18 @@ public class AddressService : IAddressService
 
     public async Task DeleteAddressAsync(AddressDelete addressDelete)
     {
-        var entity = await AddressRepository.GetByIdAsync(addressDelete.Id);
+        var entity = await AddressRepository.GetByIdAsync(addressDelete.Id, (address) => address.Employees);
+
+        if (entity == null)
+        {
+            throw new AddressNotFoundException(addressDelete.Id);
+        }
+
+        if (entity.Employees.Count > 0)
+        {
+            throw new DependentEmployeeExistException(entity.Employees);
+        }
+
         AddressRepository.Delete(entity);
         await AddressRepository.SaveChangesAsync();
     }
@@ -45,6 +57,12 @@ public class AddressService : IAddressService
     public async Task<AddressGet> GetAddressAsync(int id)
     {
         var entity = await AddressRepository.GetByIdAsync(id);
+        
+        if (entity == null)
+        {
+            throw new AddressNotFoundException(id);
+        }
+
         return Mapper.Map<AddressGet>(entity);
     }
 
@@ -57,6 +75,13 @@ public class AddressService : IAddressService
     public async Task UpdateAddressAsync(AddressUpdate addressUpdate)
     {
         await AddressUpdateValidator.ValidateAndThrowAsync(addressUpdate);
+
+        var existingAddress = await AddressRepository.GetByIdAsync(addressUpdate.Id);
+
+        if (existingAddress == null)
+        {
+            throw new AddressNotFoundException(addressUpdate.Id);
+        }
 
         var entity = Mapper.Map<Address>(addressUpdate);
         AddressRepository.Update(entity);
